** Migrated from GLI to SDL2_iamge lib for image loading * Updated TextureManager class * Refractored TextureManger class - Removed gli headers + Added SDL_image.lib & dlls for all platforms + Added SDL_image.h to SDL headers * Updated Makefile * Updated Makefile_osx * Updated the .vcxproj file * Udated DEPS_LIBS_BUILD.txt
** Migrated from .dds images to .bmp as SDL2_image does not support .dss
and .bmp is also convinient for android port (for later)
* Replaced the .dds images with .bmp in VS
* Updated the .vcxproj file
* Updated the .user.vcxproj
- Removed the .dds images
+ Added the .bmp images
* Updated the following classes: Ocean, MotorBoat, CubemapSkyModel - changed the image extension

** Changed resource path to take them in place and not from the parent directory
* Updated classes like: Main, Ocean, MotorBoat, Sky, CubeMapSkyModel, PrecomputedScatteringSkyModel, ScatteringSkyModel,
FFTNormalGradientFoldingGPUFrag, FFTNormalGradientFoldingGPUComp, FFTOceanPatchGPUComp, FFTOceanPatchGPUFrag, PostProcessingManager,
GPUComp2DIFFT, GPUFrag2DIFFT

** Migrated from std::fstream & fopen()/fscanf() to SDL2 file processing methods to maintain a cross-platform solution
* Updated FileUtils logic
* Updated Mesh & Model logic; I no longer read step by step from the .obj/.mtl files,
 but load them entirely in RAM then parse the info
* Updated TextureManager logic
* Updated ShaderManager logic
* Updated XMLParser logic

* Corrected the switch statements in all source code
* Impacted classes: Application, Ocean, FFTOceanPatchBase, FFTOceanPatchCPUFFTW, Sky, ShaderManager, FramebufferManager

** Removed the GL_EXT_texture_compression_s3tc extension
* Updated the glad_gl32.h/cpp
* Updated GLExtensionVars.h

* Minor changes in Mesh class
* Fixed the ProcessTexture() logic in Model class
+ Added override qualifer to SetFFTData() and InitFFTDdata() methods in FFTOceanPatchGPUComp.h
* Updated the GlobalConfig.xml as some adjustments were necesarry for fft & perlin noise blending
* Rearranged platform defines in AppConfig.h and SDLConfig.h
* Corrected boat.obj file as there were some unnecessary spaces
- Removed unused m_IsUnderWater bool from Application code
- Removed unused m_IsFrumstumVisible from the Projector code
- Removed obosolete geometry shader enums from ShaderManager.h
- Removed bin/win/Debug/dbghelp.dll
* Changed arguments in Camera class from float to int for:
UpdatePerspectiveProjectionMatrix(...) and GetFOV()/SetFOV() methods as well as the respective members
* Updated GlobalConfig
- Removed unnecessary methods from TextureManager
+ Added guard USE_GUI for UI
+ Added guard USE_FFTW for CPU FFT comutation
+ Added some GL version check guards

** [BUGFIX] Boat foam not visible if no kelvin wake is enabled
FIX: Use kelvin wake data for both boat foam & kelvin wake effects
+ Added computeBoatBaseEffectUV(vec3...) function in shader files
* Updated the OceanSurface.frag.glsl
* Updated the OceanSurfacePrecomputedScattering.frag.glsl
* Updated the OceanSurfaceWorldGrid.vert.glsl
* Updated the OceanSurfaceScreenGrid.vert.glsl
* Updated the MotorBoar class

** [BUGFIX] Moving too quickly when pressing SHIFT key
FIX: No longer use the updated & cached value of the key speed in Application class,
but directly use the config key speed
* Updated Application class

** [BUGFIX] No water inside the boat feature was no longer working
FIX: revised & corrected the responsible gl stencil code
* Updated Application class

** [BUGFIX] Broken visuals when using textures as depth target instead of a render buffer
FIX: corrected the wrong texture unit to which the color attachements was being bound
* Updated PostProcessingManager code - BindColorAttachmentByIndex(...)

+++ source/Camera.cpp
@@ -14,7 +14,7 @@
 Camera::Camera ( void )
 	: m_Name("Default"), m_Position(glm::vec3(0.0f)), m_Forward(glm::vec3(0.0f, 0.0f, -1.0f)),
 	  m_Right(glm::vec3(+1.0f, 0.0f, 0.0f)), m_Up(glm::vec3(0.0f, +1.0f, 0.0f)),
-	  m_Pitch(0.0f), m_Yaw(0.0f), m_FOVy(0.0f), m_InitialFOVy(0.0f), m_ZNear(0.0f), m_ZFar(0.0f),
+	  m_Pitch(0.0f), m_Yaw(0.0f), m_FOVy(0), m_InitialFOVy(0), m_ZNear(0.0f), m_ZFar(0.0f),
 	  m_UseConstraints(false)
 {
 	LOG("Camera %s successfully created!", m_Name.c_str());
@@ -23,7 +23,7 @@ Camera::Camera ( void )
 Camera::Camera ( const std::string& i_Name )
 	: m_Name(i_Name), m_Position(glm::vec3(0.0f)), m_Forward(glm::vec3(0.0f, 0.0f, -1.0f)),
 	  m_Right(glm::vec3(+1.0f, 0.0f, 0.0f)), m_Up(glm::vec3(0.0f, +1.0f, 0.0f)),
-	  m_Pitch(0.0f), m_Yaw(0.0f), m_FOVy(0.0f), m_InitialFOVy(0.0f), m_ZNear(0.0f), m_ZFar(0.0f),
+	  m_Pitch(0.0f), m_Yaw(0.0f), m_FOVy(0), m_InitialFOVy(0), m_ZNear(0.0f), m_ZFar(0.0f),
 	  m_UseConstraints(false)
 {
 	LOG("Camera %s successfully created!", m_Name.c_str());
@@ -170,7 +170,7 @@ void Camera::UpdatePerspectiveProjectionMatrix ( int i_WindowWidth, int i_Window
 
 	float correctionFactor = ComputePerspectiveProjectionCorrectionFactor();
 
-	m_Projection = glm::perspective(glm::radians(m_FOVy), i_WindowWidth /static_cast<float>(i_WindowHeight), m_ZNear * correctionFactor, m_ZFar * correctionFactor);
+	m_Projection = glm::perspective(glm::radians(static_cast<float>(m_FOVy)), i_WindowWidth /static_cast<float>(i_WindowHeight), m_ZNear * correctionFactor, m_ZFar * correctionFactor);
 
 	// update the other matrices too
 	// more info about projection matrix: http://www.songho.ca/opengl/gl_projectionmatrix.html
@@ -182,7 +182,7 @@ void Camera::UpdatePerspectiveProjectionMatrix ( int i_WindowWidth, int i_Window
 	m_InverseProjectionView = glm::inverse(m_ProjectionView);
 }
 
-void Camera::UpdatePerspectiveProjectionMatrix ( float i_Fovy, float i_Aspect, float i_ZNear, float i_ZFar )
+void Camera::UpdatePerspectiveProjectionMatrix ( int i_Fovy, float i_Aspect, float i_ZNear, float i_ZFar )
 {
 	assert(i_Fovy > 0);
 	assert(i_Aspect > 0);
@@ -191,7 +191,7 @@ void Camera::UpdatePerspectiveProjectionMatrix ( float i_Fovy, float i_Aspect, f
 
 	float correctionFactor = ComputePerspectiveProjectionCorrectionFactor();
 
-	m_Projection = glm::perspective(glm::radians(i_Fovy), i_Aspect, i_ZNear * correctionFactor, i_ZFar * correctionFactor);
+	m_Projection = glm::perspective(glm::radians(static_cast<float>(i_Fovy)), i_Aspect, i_ZNear * correctionFactor, i_ZFar * correctionFactor);
 
 	// update the other matrices too
 	// more info about projection matrix: http://www.songho.ca/opengl/gl_projectionmatrix.html
@@ -301,7 +301,7 @@ float Camera::GetYaw ( void ) const
 	return m_Yaw;
 }
 
-float Camera::GetFOV ( void ) const
+int Camera::GetFOV ( void ) const
 {
 	return m_FOVy;
 }
@@ -361,7 +361,7 @@ void Camera::SetUp ( const glm::vec3& i_Up )
 	m_Up = i_Up;
 }
 
-void Camera::SetFOV ( float i_FOV )
+void Camera::SetFOV ( int i_FOV )
 {
 	m_FOVy = i_FOV;
 }
